Changed stability to double

+++ src/Gateway/Models/Broadcast.cs
@@ -12,7 +12,7 @@ namespace Gateway.Models
         public string Id { get; set; }
         public double[] Categories { get; set; }
         public int Bitrate { get; set; }
-        public float Stability { get; set; }
+        public double Stability { get; set; }
         public Location Location { get; set; }
         public DateTime Activity { get; set; }
         public string AccountId { get; set; }
